requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
iniconfig==2.0.0
kombu==5.3.5
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-celery==0.0.0
pytest-django==4.8.0
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.4
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
